实现告警信息的处置功能；

+++ core/src/main/java/com/mingshi/skyflying/domain/MsAlarmInformationDo.java
@@ -9,7 +9,7 @@ public class MsAlarmInformationDo {
 
     private Integer matchRuleId;
 
-    private Byte isDelete;
+    private Integer isDelete;
 
     private Date gmtCreate;
